backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Data validation
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.2.0

# Database
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Task queue
celery==5.4.0
redis==5.0.7
kombu==5.3.7

# Image / OCR
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
pdf2image==1.17.0
paddleocr==2.7.3
paddlepaddle==2.6.2
azure-ai-formrecognizer==3.3.3
azure-storage-blob==12.20.0

# Utilities
httpx==0.27.0
tenacity==8.4.2
python-dotenv==1.0.1
structlog==24.2.0
loguru==0.7.2

# Observability / resilience
prometheus-fastapi-instrumentator==7.0.0
pybreaker==1.2.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0
factory-boy==3.3.0
